requirements (4).txt
streamlit
google-cloud-texttospeech
python-pptx
moviepy
imageio[ffmpeg]  # moviepy에서 ffmpeg를 사용하기 위해 추가
